session based oauth

+++ src/main/java/com/sherpa/interviewsherpa/flow/domain/viewport/Viewport.java
@@ -3,5 +3,5 @@ package com.sherpa.interviewsherpa.flow.domain.viewport;
 import lombok.Builder;
 
 @Builder
-public record Viewport(int x, int y, int zoom) {
+public record Viewport(Double x, Double y, Double zoom) {
 }
